Issue #285 express-session free support
We used to save state and nonce in session, so session is always required. Now we can encrypt and decrypt
state and nonce, and put them into cookie, so session is no longer mandatory.

We provided two more options: useCookieInsteadOfSession, cookieEncryptionKeys. To use cookie, first set the first option to true.
For the second option, provide an array of { key: '...', iv: '...' } items. For each item, key is any string
of size 32, and iv is any string of size 12. We will use the first item (with AES-256-GCM algorithm) to encrypt cookie, but we try every
item to decrypt cookie, so you can easily do key rollover.